Code review amendments - rename and further organise file structure
