+++ server/middleware/getPhotoOfTheDay.js
@@ -1,5 +1,5 @@
 const fetch = require("node-fetch");
-const API_KEY = require("../../apiKey");
+const API_KEY = require("../apiKey");
 const BASE_URL = "https://api.nasa.gov/planetary/apod";
 
 const checkStatus = (response) => {
@@ -15,6 +15,7 @@ const getPhotoOfTheDay = async (request, response, next) => {
   try {
     const apodResponse = await fetch(url).then(checkStatus);
     const apodJson = await apodResponse.json();
+    console.log(apodJson);
     response.json(apodJson);
   } catch (error) {
     return response.status(500).json({ error: error.toString() });
